# swaggerui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Package main",
    "version": "0.0.1"
  },
  "host": "localhost:10000",
  "basePath": "/",
  "paths": {
...
    "/api/proxy/group": {
      "get": {
        "description": "Returns a list of all ProxyGroups",
        "tags": [
          "ProxyGroup"
        ],
        "operationId": "GetAllProxyGroupsEndpoint",
        "responses": {
          "200": {
            "description": "ProxyGroup response",
            "schema": {
              "$ref": "#/responses/ProxyGroupResponseSwagger"
            }
          }
        }
      },
      "post": {
        "description": "Creates a ProxyGroup in the database",
        "tags": [
          "ProxyGroup"
        ],
        "operationId": "CreateProxyGroupEndpoint",
        "parameters": [
          {
            "description": "ProxyGroup details",
            "name": "ProxyGroupDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateProxyGroupRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProxyGroup response",
            "schema": {
              "$ref": "#/responses/ProxyGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/proxy/group/{GroupID}": {
      "get": {
        "tags": [
          "ProxyGroup"
        ],
        "summary": "Returns the ProxyGroup with GroupID {GroupID}.",
        "operationId": "GetProxyGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProxyGroup to retrieve",
            "name": "GroupID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ProxyGroup response",
            "schema": {
              "$ref": "#/responses/ProxyGroupResponseSwagger"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProxyGroup"
        ],
        "summary": "Updates and returns the ProxyGroup with GroupID {GroupID}.",
        "operationId": "UpdateProxyGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProxyGroup to update",
            "name": "GroupID",
            "in": "path"
          },
          {
            "description": "Details to update ProxyGroup with",
            "name": "ProxyGroupDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateProxyGroupRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProxyGroup response",
            "schema": {
              "$ref": "#/responses/ProxyGroupResponseSwagger"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProxyGroup"
        ],
        "summary": "Deletes and returns the ProxyGroup with GroupID {GroupID}.",
        "operationId": "DeleteProxyGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProxyGroup to retrieve",
            "name": "GroupID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ProxyGroup response",
            "schema": {
              "$ref": "#/responses/ProxyGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/proxy/group/{GroupID}/clone": {
      "post": {
        "tags": [
          "ProxyGroup"
        ],
        "summary": "Clones the ProxyGroup with GroupID {GroupID} and returns the clone.",
        "operationId": "CloneProxyGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProxyGroup to clone",
            "name": "GroupID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "ProxyGroup response",
            "schema": {
              "$ref": "#/responses/ProxyGroupResponseSwagger"
            }
          }
        }
      }
    },
...
  }
}